chore(apps/api): fix build
