+++ apps/api/src/services/flts.ts
@@ -26,10 +26,12 @@ const queryFuseLeveragedTokens = gql`
             collateral {
                 name
                 symbol
+                decimals
             }
             debt {
                 name
                 symbol
+                decimals
             }
         }
     }
@@ -128,25 +130,6 @@ const queryFuseLeveragedTokenBackingsBySymbol = gql`
     }
 `;
 
-const queryUserPositionById = gql`
-    query getOpenPosition($id: ID!) {
-        position: openPosition(id: $id) {
-            flt {
-                price: fltHourData(
-                    first: 1
-                    orderBy: periodStartUnix
-                    orderDirection: desc
-                ) {
-                    timestamp: periodStartUnix
-                    close
-                }
-            }
-            amount
-            amountUSD
-        }
-    }
-`;
-
 // prettier-ignore
 const BSC_GRAPH = "https://api.thegraph.com/subgraphs/name/risedle/risedle-flt-bsc";
 
@@ -196,11 +179,13 @@ function getFuseLeveragedTokenInfo(flt: any): FuseLeveragedTokenInfo {
             name: flt.collateral.name,
             symbol: flt.collateral.symbol,
             amount: collateralPerShare,
+            decimals: parseInt(flt.collateral.decimals),
         },
         debt: {
             name: flt.debt.name,
             symbol: flt.debt.symbol,
             amount: debtPerShare,
+            decimals: parseInt(flt.debt.decimals),
         },
         totalCollateral: totalCollateral,
         totalDebt: totalDebt,
@@ -283,13 +268,6 @@ interface FuseLeveragedTokenBacking {
 
 interface FuseLeveragedTokenBackings {
     backings: Array<FuseLeveragedTokenBacking>;
-}
-
-interface FuseLeveragedTokenUserPosition {
-    balance: number;
-    usd: number; // usd value of current balance
-    pnlUSD: number; // open p/l in USD
-    pnlPercent: number; // open p/l in percentage
 }
 
 /**
@@ -412,35 +390,11 @@ export async function getFuseLeveragedTokenBackingsBySymbol(
     return { backings };
 }
 
-/**
- * Get user position by id
- */
-export async function getUserPositionById(
-    chainId: ChainId,
-    positionId: string
-): Promise<FuseLeveragedTokenUserPosition | undefined> {
-    // Get data from the graph
-    const endpoint = getGraphEndpointByChainId(chainId);
-    const filter = positionId.toLowerCase();
-    const data = await grequest(endpoint, queryUserPositionById, {
-        id: filter,
-    });
-    if (data.position == null) return undefined;
-    const balance = parseFloat(data.position.amount);
-    const price = parseFloat(data.position.flt.price[0].close);
-    const usd = balance * price;
-    const principal = parseFloat(data.position.amountUSD);
-    const pnlUSD = usd - principal;
-    const pnlPercent = (pnlUSD / principal) * 100;
-    return { balance, usd, pnlUSD, pnlPercent };
-}
-
 const flts = {
     getFuseLeveragedTokensByChainId,
     getFuseLeveragedTokenChartsBySymbol,
     getFuseLeveragedTokenSwapsBySymbol,
     getFuseLeveragedTokenBackingsBySymbol,
-    getUserPositionById,
 };
 
 export default flts;
